backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
supabase==2.0.2
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# AI/ML (OpenAI)
openai==1.3.7

# Utilities
pydantic==2.5.0
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0.post1

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Platform Integrations
slack-sdk==3.26.0

# Monitoring & Logging
structlog==23.2.0